requirements.txt
Django==6.0.3
djangorestframework==3.17.0
django-cors-headers==4.9.0
Pillow==12.1.1
python-decouple==3.8
django-filter==25.2
drf-yasg==1.21.15
celery==5.6.2
redis==7.3.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
pyyaml==6.0.3
requests==2.31.0
